requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
azure-identity==1.15.0
azure-kusto-data==4.5.1
jinja2==3.1.2
python-multipart==0.0.6
pydantic-settings==2.1.0